Separate frequent subgraph code in a class
