+++ src/fixrgraphiso/CMakeLists.txt
@@ -43,6 +43,8 @@ target_link_libraries(frequentitemsets
 
 add_executable(frequentsubgraphs
    frequentSubgraphsMain.cpp
+   frequentSubgraphs.cpp
+   frequentSubgraphs.h
    acdfgBin.cpp
    isomorphismClass.cpp
    acdfg.cpp
@@ -55,6 +57,8 @@ target_compile_features(frequentsubgraphs PRIVATE cxx_range_for)
 set(CMAKE_MACOSX_RPATH 1)
 add_library(frequentsubgraphs_library SHARED
    frequentSubgraphsMain.cpp
+   frequentSubgraphs.cpp
+   frequentSubgraphs.h
    acdfgBin.cpp
    isomorphismClass.cpp
    acdfg.cpp
